# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FitMeet",
    "description": "Backend do projeto FitMeet feito como desafio do Bootcamp da Sysmap",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/activities/types": {
      "get": {
        "summary": "Listar Tipos de atividades disponíveis",
        "tags": ["Atividades"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de atividades",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "name": { "type": "string" },
                      "description": { "type": "string" },
                      "image": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activities": {
      "get": {
        "summary": "Listar atividades com paginação, fitro por tipo e ordenação",
        "tags": ["Atividades"],
        "security": [{ "bearerAuth": [] }],
        "description": "Endpoint para listar de forma paginada as atividades registradas. Parametros opcionais: type(filtrar por tipo especifico de atividades), orderBy(campo pelo qual ordenar as atividades) e order ('asc'=crescente, 'desc': = decrescente).",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "pagesize",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "typeId",
            "required": false,
            "schema": {
              "type": "uuid"
            }
          },
          {
            "in": "query",
            "name": "orderBy",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de atividades",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": { "type": "integer", "example": 0 },
                    "pageSize": { "type": "integer", "example": 0 },
                    "totalActivities": { "type": "integer", "example": 0 },
                    "totalPages": { "type": "integer", "example": 0 },
                    "previous": { "type": "integer", "example": 0 },
                    "next": { "type": "integer", "example": 0 },
                    "activities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "title": { "type": "string" },
                          "description": { "type": "string" },
                          "type": { "type": "string" },
                          "image": { "type": "string" },
                          "confirmationCode": { "type": "string" },
                          "participantCount": { "type": "integer" },
                          "address": {
                            "type": "object",
                            "properties": {
                              "latitude": { "type": "number" },
                              "longitude": { "type": "number" }
                            }
                          },
                          "scheduledDate": { "type": "string", "format": "date-time" },
                          "createdAt": { "type": "string", "format": "date-time" },
                          "completedAt": { "type": "string", "format": "date-time" },
                          "private": { "type": "boolean" },
                          "creator": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string", "format": "uuid" },
                              "name": { "type": "string" },
                              "avatar": { "type": "string" }
                            }
                          },
                          "userSubscriptionStatus": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activities/all": {
      "get": {
        "summary": "Listar todas atividades com paginação, fitro por tipo e ordenação",
        "tags": ["Atividades"],
        "security": [{ "bearerAuth": [] }],
        "description": "Endpoint para listar de forma paginada as atividades registradas. Parametros opcionais: type(filtrar por tipo especifico de atividades), orderBy(campo pelo qual ordenar as atividades) e order ('asc'=crescente, 'desc': = decrescente).",
        "parameters": [
          {
            "in": "query",
            "name": "typeId",
            "required": false,
            "schema": {
              "type": "uuid"
            }
          },
          {
            "in": "query",
            "name": "orderBy",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de atividades",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",

                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "title": { "type": "string" },
                      "description": { "type": "string" },
                      "type": { "type": "string" },
                      "image": { "type": "string" },
                      "confirmationCode": { "type": "string" },
                      "participantCount": { "type": "integer" },
                      "address": {
                        "type": "object",
                        "properties": {
                          "latitude": { "type": "number" },
                          "longitude": { "type": "number" }
                        }
                      },
                      "scheduledDate": { "type": "string", "format": "date-time" },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "completedAt": { "type": "string", "format": "date-time" },
                      "private": { "type": "boolean" },
                      "creator": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": { "type": "string" },
                          "avatar": { "type": "string" }
                        }
                      },
                      "userSubscriptionStatus": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activities/user/creator": {
      "get": {
        "summary": "Listar  atividades criada pelo usuario",
        "tags": ["Atividades"],
        "security": [{ "bearerAuth": [] }],
        "description": "Endpoint para listar de forma paginada todas as atividades criadas pelo usuario logado.",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "pagesize",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de atividades",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": { "type": "integer", "example": 0 },
                    "pageSize": { "type": "integer", "example": 0 },
                    "totalActivities": { "type": "integer", "example": 0 },
                    "totalPages": { "type": "integer", "example": 0 },
                    "previous": { "type": "integer", "example": 0 },
                    "next": { "type": "integer", "example": 0 },
                    "activities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "title": { "type": "string" },
                          "description": { "type": "string" },
                          "type": { "type": "string" },
                          "image": { "type": "string" },
                          "confirmationCode": { "type": "string" },
                          "participantCount": { "type": "integer" },
                          "address": {
                            "type": "object",
                            "properties": {
                              "latitude": { "type": "number" },
                              "longitude": { "type": "number" }
                            }
                          },
                          "scheduledDate": { "type": "string", "format": "date-time" },
                          "createdAt": { "type": "string", "format": "date-time" },
                          "completedAt": { "type": "string", "format": "date-time" },
                          "private": { "type": "boolean" },
                          "creator": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string", "format": "uuid" },
                              "name": { "type": "string" },
                              "avatar": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activities/user/creator/all": {
      "get": {
        "summary": "Listar todas atividades criada pelo usuario.",
        "tags": ["Atividades"],
        "security": [{ "bearerAuth": [] }],
        "description": "Endpoint para listar todas as atividades criadas pelo usuario logado.",
        "responses": {
          "200": {
            "description": "Lista de atividades",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "title": { "type": "string" },
                      "description": { "type": "string" },
                      "type": { "type": "string" },
                      "image": { "type": "string" },
                      "confirmationCode": { "type": "string" },
                      "participantCount": { "type": "integer" },
                      "address": {
                        "type": "object",
                        "properties": {
                          "latitude": { "type": "number" },
                          "longitude": { "type": "number" }
                        }
                      },
                      "scheduledDate": { "type": "string", "format": "date-time" },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "completedAt": { "type": "string", "format": "date-time" },
                      "private": { "type": "boolean" },
                      "creator": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": { "type": "string" },
                          "avatar": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/activities/user/participant/all": {
      "get": {
        "summary": "Buscar todas as atividades que o usuario se inscreveu",
        "tags": ["Atividades"],
        "security": [{ "bearerAuth": [] }],
        "description": "Endpoint para listar todas as atividades em que o usuario logado se inscreveu.",
        "responses": {
          "200": {
            "description": "Lista de atividades",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",

                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "title": { "type": "string" },
                      "description": { "type": "string" },
                      "type": { "type": "string" },
                      "image": { "type": "string" },
                      "confirmationCode": { "type": "string" },
                      "participantCount": { "type": "integer" },
                      "address": {
                        "type": "object",
                        "properties": {
                          "latitude": { "type": "number" },
                          "longitude": { "type": "number" }
                        }
                      },
                      "scheduledDate": { "type": "string", "format": "date-time" },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "completedAt": { "type": "string", "format": "date-time" },
                      "private": { "type": "boolean" },
                      "creator": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": { "type": "string" },
                          "avatar": { "type": "string" }
                        }
                      },
                      "userSubscriptionStatus": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}